test: move metadata test into integration tests
