+++ .github/workflows/ci.yml
@@ -264,5 +264,4 @@ jobs:
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
         run: |
-          cargo test --no-default-features --features chain --test chain
-          cargo test -p pop-common --test metadata
+          cargo test --no-default-features --features chain --test chain --test metadata
