+++ crates/pop-chains/src/call/metadata/params.rs
@@ -158,99 +158,3 @@ fn type_to_param(name: &str, registry: &PortableRegistry, type_id: u32) -> Resul
 		}
 	}
 }
-
-#[cfg(test)]
-mod tests {
-	use super::*;
-	use crate::{call::tests::POP_NETWORK_TESTNET_URL, set_up_client};
-	use anyhow::Result;
-
-	// TODO: Integration test, parsing metadata
-	#[tokio::test]
-	async fn field_to_param_works() -> Result<()> {
-		let client = set_up_client(POP_NETWORK_TESTNET_URL).await?;
-		let metadata = client.metadata();
-		// Test a supported dispatchable function.
-		let function = metadata
-			.pallet_by_name("Balances")
-			.unwrap()
-			.call_variant_by_name("force_transfer")
-			.unwrap();
-		let mut params = Vec::new();
-		for field in &function.fields {
-			params.push(field_to_param(&metadata, field)?)
-		}
-		assert_eq!(params.len(), 3);
-		assert_eq!(params.first().unwrap().name, "source");
-		assert_eq!(params.first().unwrap().type_name, "MultiAddress<AccountId32 ([u8;32]),()>: Id(AccountId32 ([u8;32])), Index(Compact<()>), Raw([u8]), Address32([u8;32]), Address20([u8;20])");
-		assert_eq!(params.first().unwrap().sub_params.len(), 5);
-		assert_eq!(params.first().unwrap().sub_params.first().unwrap().name, "Id");
-		assert_eq!(params.first().unwrap().sub_params.first().unwrap().type_name, "");
-		assert_eq!(
-			params
-				.first()
-				.unwrap()
-				.sub_params
-				.first()
-				.unwrap()
-				.sub_params
-				.first()
-				.unwrap()
-				.name,
-			"Id"
-		);
-		assert_eq!(
-			params
-				.first()
-				.unwrap()
-				.sub_params
-				.first()
-				.unwrap()
-				.sub_params
-				.first()
-				.unwrap()
-				.type_name,
-			"AccountId32 ([u8;32])"
-		);
-		// Test some dispatchable functions that are not supported.
-		let function =
-			metadata.pallet_by_name("Sudo").unwrap().call_variant_by_name("sudo").unwrap();
-		assert!(matches!(
-			field_to_param(&metadata, &function.fields.first().unwrap()),
-			Err(Error::FunctionNotSupported)
-		));
-		let function = metadata
-			.pallet_by_name("Utility")
-			.unwrap()
-			.call_variant_by_name("batch")
-			.unwrap();
-		assert!(matches!(
-			field_to_param(&metadata, &function.fields.first().unwrap()),
-			Err(Error::FunctionNotSupported)
-		));
-		let function = metadata
-			.pallet_by_name("PolkadotXcm")
-			.unwrap()
-			.call_variant_by_name("execute")
-			.unwrap();
-		assert!(matches!(
-			field_to_param(&metadata, &function.fields.first().unwrap()),
-			Err(Error::FunctionNotSupported)
-		));
-		// TODO: Use the Pop Network endpoint once the mainnet (or an equivalent testnet with the
-		// same runtime) is available.
-		let client = set_up_client("wss://mythos.ibp.network").await?;
-		let metadata = client.metadata();
-		let function = metadata
-			.pallet_by_name("Council")
-			.unwrap()
-			.call_variant_by_name("execute")
-			.unwrap();
-		assert!(matches!(
-			field_to_param(&metadata, &function.fields.first().unwrap()),
-			Err(Error::FunctionNotSupported)
-		));
-
-		Ok(())
-	}
-}
